+++ modules/tinymce/src/core/test/ts/webdriver/PageUpDownKeysTest.ts
@@ -1,6 +1,6 @@
-import { Keys, RealKeys } from '@ephox/agar';
-import { context, describe, it } from '@ephox/bedrock-client';
-import { TinyAssertions, TinyContentActions, TinyHooks, TinySelections } from '@ephox/mcagar';
+import { RealKeys } from '@ephox/agar';
+import { before, context, describe, it } from '@ephox/bedrock-client';
+import { TinyAssertions, TinyHooks, TinySelections } from '@ephox/mcagar';
 import { PlatformDetection } from '@ephox/sand';
 
 import Editor from 'tinymce/core/api/Editor';
@@ -13,89 +13,85 @@ describe('webdriver.tinymce.core.keyboard.PageUpDownKeyTest', () => {
     indent: false
   }, [ Theme ], true);
   const platform = PlatformDetection.detect();
+  const supportsPageUpDown = !(platform.os.isOSX() || platform.os.isWindows() && platform.browser.isFirefox());
 
-  // It necessary to add a different assertion for Firefox because it does not move the caret start/end on OS windows and Mac,
-  // Linux Firefox could be an exception, nevertheless we does not override the default behavior.
+  // It's necessary to skip tests for Mac and Windows Firefox as they don't move the selection when pressing page up/down
+  before(function () {
+    if (!supportsPageUpDown) {
+      this.skip();
+    }
+  });
+
   context('Page Up', () => {
-    it('TINY-4612: caret should be placed at the start of the line if the first child is an inline element', async () => {
+    it('TINY-4612: caret should be placed at the start of the line', async () => {
       const editor = hook.editor();
       editor.setContent('<p><a href="google.com">link</a>text</p>');
       TinySelections.setCursor(editor, [ 0, 1 ], 2);
       await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.combo({}, 'PageUp') ]);
-      TinyContentActions.keystroke(editor, Keys.pageUp());
 
       TinyAssertions.assertCursor(editor, [ 0 ], 0);
     });
 
-    it('TINY-4612: macOS wont move the caret if it is not at inline element at the start of the line, whereas other OSs do', async () => {
+    it('TINY-4612: caret should be placed out of the line element', async () => {
       const editor = hook.editor();
-      editor.setContent('<p>text<a href="google.com">link2</a>text</p>');
-      TinySelections.setCursor(editor, [ 0, 2 ], 0);
+      editor.setContent('<p><a href="google.com">link</a></p>');
+      TinySelections.setCursor(editor, [ 0, 0, 0 ], 2);
       await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.combo({}, 'PageUp') ]);
-      TinyContentActions.keystroke(editor, Keys.pageUp());
 
-      if (platform.os.isOSX() || platform.browser.isFirefox()) {
-        TinyAssertions.assertCursor(editor, [ 0, 2 ], 0);
-      } else {
-        TinyAssertions.assertCursor(editor, [ 0, 0 ], 0);
-      }
+      TinyAssertions.assertCursor(editor, [ 0 ], 0);
     });
 
-    it('TINY-4612: inline_boundaries: false macOS doesnt move, whereas other OSs do', async () => {
+    it('TINY-4612: "inline_boundaries: false" macOS doesnt move, whereas other OSs do', async () => {
       const editor = hook.editor();
+      editor.settings.height = 400;
       editor.settings.inline_boundaries = false;
-      editor.setContent('<p>test<a href="google.com">link</a>text</p>');
-      TinySelections.setCursor(editor, [ 0, 2 ], 2);
+      editor.setContent('<p><a href="google.com">link</a>text</p>');
+      TinySelections.setCursor(editor, [ 0, 1 ], 4);
       await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.combo({}, 'PageUp') ]);
-      TinyContentActions.keystroke(editor, Keys.pageUp());
 
-      if (platform.os.isOSX() || platform.browser.isFirefox()) {
-        TinyAssertions.assertCursor(editor, [ 0, 2 ], 2);
+      if (platform.browser.isIE()) {
+        TinyAssertions.assertCursor(editor, [ 0 ], 0);
       } else {
-        TinyAssertions.assertCursor(editor, [ 0, 0 ], 0);
+        TinyAssertions.assertCursor(editor, [ 0, 0, 0 ], 0);
       }
+
       delete editor.settings.inline_boundaries;
     });
   });
 
   context('Page Down', () => {
-    it('TINY-4612: caret should be placed at the end of the line if the last child is an inline element', async () => {
+    it('TINY-4612: caret should be placed at the end of the line', async () => {
       const editor = hook.editor();
       editor.setContent('<p>text<a href="google.com">link</a></p>');
       TinySelections.setCursor(editor, [ 0, 0 ], 0);
       await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.combo({}, 'PageDown') ]);
-      TinyContentActions.keystroke(editor, Keys.pageDown());
 
       TinyAssertions.assertCursor(editor, [ 0 ], 2);
     });
 
-    it('TINY-4612: macOS wont move the caret if it is not at inline element at the end of the line, whereas other OSs do', async () => {
+    it('TINY-4612: caret should be placed out of the line element', async () => {
       const editor = hook.editor();
-      editor.setContent('<p>test<a href="google.com">link</a>text</p>');
-      TinySelections.setCursor(editor, [ 0, 0 ], 0);
+      editor.setContent('<p><a href="google.com">link</a></p>');
+      TinySelections.setCursor(editor, [ 0, 0, 0 ], 2);
       await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.combo({}, 'PageDown') ]);
-      TinyContentActions.keystroke(editor, Keys.pageDown());
 
-      if (platform.os.isOSX() || platform.browser.isFirefox()) {
-        TinyAssertions.assertCursor(editor, [ 0, 0 ], 0);
-      } else {
-        TinyAssertions.assertCursor(editor, [ 0, 2 ], 4);
-      }
+      TinyAssertions.assertCursor(editor, [ 0 ], 1);
     });
 
-    it('TINY-4612: inline_boundaries: false macOS doesnt move, whereas other OSs do', async () => {
+    it('TINY-4612: "inline_boundaries: false" cursor does not move outside anchor', async () => {
       const editor = hook.editor();
+      editor.settings.height = 400;
       editor.settings.inline_boundaries = false;
-      editor.setContent('<p>test<a href="google.com">link</a>text</p>');
+      editor.setContent('<p>test<a href="google.com">link</a></p>');
       TinySelections.setCursor(editor, [ 0, 0 ], 0);
       await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.combo({}, 'PageDown') ]);
-      TinyContentActions.keystroke(editor, Keys.pageDown());
 
-      if (platform.os.isOSX() || platform.browser.isFirefox()) {
-        TinyAssertions.assertCursor(editor, [ 0, 0 ], 0);
+      if (platform.browser.isIE()) {
+        TinyAssertions.assertCursor(editor, [ 0 ], 2);
       } else {
-        TinyAssertions.assertCursor(editor, [ 0, 2 ], 4);
+        TinyAssertions.assertCursor(editor, [ 0, 1, 0 ], 4);
       }
+
       delete editor.settings.inline_boundaries;
     });
   });